Extract matching system fonts.
